requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
pypdf
fastapi
uvicorn
streamlit
requests